+++ iree_tests/configs/config_onnx_cpu_llvm_sync.json
@@ -30,8 +30,6 @@
     "test_ai_onnx_ml_label_encoder_string_int_no_default",
     "test_ai_onnx_ml_label_encoder_tensor_value_only_mapping",
     "test_averagepool_2d_dilations",
-    "test_averagepool_2d_pads",
-    "test_averagepool_2d_precomputed_pads",
     "test_averagepool_2d_precomputed_same_upper",
     "test_averagepool_2d_same_lower",
     "test_averagepool_2d_same_upper",
@@ -134,8 +132,6 @@
     "test_dequantizelinear_e4m3fn_float16",
     "test_dequantizelinear_e4m3fn_zero_point",
     "test_dequantizelinear_e5m2",
-    "test_dequantizelinear_int16",
-    "test_dequantizelinear_uint16",
     "test_det_2d",
     "test_det_nd",
     "test_dft",
@@ -260,7 +256,6 @@
     "test_quantizelinear_blocked",
     "test_quantizelinear_e4m3fn",
     "test_quantizelinear_e5m2",
-    "test_quantizelinear_int16",
     "test_quantizelinear_uint16",
     "test_range_float_type_positive_delta_expanded",
     "test_range_int32_type_negative_delta_expanded",
@@ -373,12 +368,9 @@
     "test_resize_downsample_scales_cubic_align_corners",
     "test_resize_downsample_scales_cubic_antialias",
     "test_resize_downsample_scales_linear_antialias",
-    "test_resize_downsample_scales_linear_half_pixel_symmetric",
-    "test_resize_downsample_scales_nearest",
     "test_resize_downsample_sizes_cubic",
     "test_resize_downsample_sizes_cubic_antialias",
     "test_resize_downsample_sizes_linear_antialias",
-    "test_resize_downsample_sizes_nearest",
     "test_resize_downsample_sizes_nearest_not_larger",
     "test_resize_downsample_sizes_nearest_not_smaller",
     "test_resize_tf_crop_and_resize",
@@ -388,17 +380,13 @@
     "test_resize_upsample_scales_cubic_A_n0p5_exclude_outside",
     "test_resize_upsample_scales_cubic_align_corners",
     "test_resize_upsample_scales_cubic_asymmetric",
-    "test_resize_upsample_scales_linear_half_pixel_symmetric",
-    "test_resize_upsample_scales_nearest",
     "test_resize_upsample_scales_nearest_axes_2_3",
     "test_resize_upsample_scales_nearest_axes_3_2",
     "test_resize_upsample_sizes_cubic",
-    "test_resize_upsample_sizes_nearest",
     "test_resize_upsample_sizes_nearest_axes_2_3",
     "test_resize_upsample_sizes_nearest_axes_3_2",
-    "test_resize_upsample_sizes_nearest_ceil_half_pixel",
+    "test_resize_upsample_sizes_nearest_floor_align_corners",
     "test_resize_upsample_sizes_nearest_not_larger",
-    "test_resize_upsample_sizes_nearest_round_prefer_ceil_asymmetric",
     "test_reversesequence_batch",
     "test_reversesequence_time",
     "test_rnn_seq_length",
@@ -537,12 +525,10 @@
     "test_unique_sorted_with_axis_3d",
     "test_unique_sorted_with_negative_axis",
     "test_unique_sorted_without_axis",
-    "test_upsample_nearest",
     "test_wrap_pad"
   ],
   "expected_run_failures": [
     "test_asin",
-    "test_averagepool_2d_ceil",
     "test_bernoulli",
     "test_bernoulli_double",
     "test_bernoulli_double_expanded",
@@ -608,7 +594,6 @@
     "test_reduce_min_empty_set",
     "test_reduce_sum_empty_set_non_reduced_axis_zero",
     "test_resize_downsample_scales_linear_align_corners",
-    "test_resize_upsample_sizes_nearest_floor_align_corners",
     "test_scatter_elements_with_negative_indices",
     "test_sce_mean_no_weight_ii",
     "test_sce_mean_no_weight_ii_log_prob",
