+++ iree_tests/configs/config_sdxl_scheduled_unet_cpu_llvm_task.json
@@ -18,7 +18,5 @@
     "skip_compile_tests": [],
     "skip_run_tests": [],
     "expected_compile_failures": [],
-    "expected_run_failures": [
-      "sdxl-scheduled-unet-3-tank",
-    ]
+    "expected_run_failures": []
 }
