+++ iree_tests/requirements.txt
@@ -5,6 +5,10 @@ pytest-retry
 pytest-timeout
 pytest-xdist
 pyjson5
+pytest-check
 
 # Remote file downloading
 azure-storage-blob
+
+# Markdown table generator
+tabulate
